+++ libraries/AP_Motors/AP_MotorsMulticopter.h
@@ -38,7 +38,7 @@ public:
 
     // output - sends commands to the motors
     virtual void        output();
-    virtual void        output(uint16_t &var1);
+    virtual void        output(float &var1);
 
     // output_min - sends minimum values out to the motors
     void                output_min();
@@ -110,7 +110,7 @@ protected:
 
     // mathaus
     // output_to_motors - sends commands to the motors
-    virtual void        output_to_motors(uint16_t &var) = 0;
+    virtual void        output_to_motors(float &var) = 0;
 
     // update the throttle input filter
     virtual void        update_throttle_filter();
